Created function package under core.util and moved the Function inside it.
