+++ sdk/core/azure-core/src/main/java/com/azure/android/core/util/function/Function.java
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-package com.azure.android.core.util;
+package com.azure.android.core.util.function;
 
 /**
  * Simple interface make code compatible with Java 7
